requirements.txt
streamlit
google-generativeai
PyPDF2
nltk
gTTS
pygame
groq